Bug: Fix Tutorial Tom
Fixed Softlock with Tutorial Tom[And some intro quests], the attempt was to use quests.txt, but I don't honestly see a good reason to place that into phasefiles anyways since it's not a file a user should change ever.

I changed it to use QuestScripts.py instead, changing the ADD_INVENTORY to add 1 instead inorder to fix the bug of fake adding 2 of each gag instead of 1.